[fix] Adjust distance and status values in tests to match mock expectations
